+++ default_www/backend/modules/locale/installer/install.php
@@ -45,6 +45,8 @@ class LocaleInstall extends ModuleInstaller
 		$this->setActionRights(1, 'locale', 'index');
 		$this->setActionRights(1, 'locale', 'mass_action');
 		$this->setActionRights(1, 'locale', 'save_translation');
+		$this->setActionRights(1, 'locale', 'export');
+		$this->setActionRights(1, 'locale', 'import');
 	}
 }