+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+2.x.x (2011-xx-xx)
+--
+Improvements:
+
+* Profiles: added profiles module to handle onsite (frontend) profiles.
+* Groups: addes groups module to handle backend user privileges.
+* Locale: added quick-edit.
+
 2.3.1 (2011-06-14)
 --
 Bugfixes: